% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLGdata-package.R
\docType{package}
\name{MLGdata-package}
\alias{MLGdata}
\alias{MLGdata-package}
\title{MLGdata: Datasets for Used with Salvan, Sartori and Pace (2020)}
\description{
The package contains the datasets for use with the book Salvan, Sartori and Pace (2020, ISBN:978-88-470-4002-1) "Modelli Lineari Generalizzati".
}
\keyword{internal}
